Update datasource.py
Updated code to address this helper by chatgpt : Make sure that you are inserting the new column element in the correct location in the XML tree. The column element should be a child element of the datasource element, and should come after any dimension or measure elements that are already present in the XML tree.